music_player/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
black==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.4
flake8==7.1.2
frozenlist==1.5.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
propcache==0.2.1
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
twilio==9.4.3
tzdata==2024.2
urllib3==2.3.0
yarl==1.18.3
